+++ driver-core/src/main/java/com/datastax/driver/core/Session.java
@@ -21,6 +21,9 @@ import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.ListeningExecutorService;
+import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.Uninterruptibles;
 
 import com.datastax.driver.core.exceptions.*;
@@ -258,7 +261,6 @@ public class Session {
         final Cluster cluster;
 
         final ConcurrentMap<Host, HostConnectionPool> pools;
-        final LoadBalancingPolicy loadBalancer;
 
         final HostConnectionPool.PoolState poolsState;
 
@@ -268,7 +270,6 @@ public class Session {
             this.cluster = cluster;
 
             this.pools = new ConcurrentHashMap<Host, HostConnectionPool>(hosts.size());
-            this.loadBalancer = cluster.manager.configuration.getPolicies().getLoadBalancingPolicy();
             this.poolsState = new HostConnectionPool.PoolState();
 
             // Create pool to initial nodes (and wait for them to be created)
@@ -298,7 +299,15 @@ public class Session {
             return cluster.manager.configuration;
         }
 
-        public ExecutorService executor() {
+        LoadBalancingPolicy loadBalancingPolicy() {
+            return cluster.manager.loadBalancingPolicy();
+        }
+
+        ReconnectionPolicy reconnectionPolicy() {
+            return cluster.manager.reconnectionPolicy();
+        }
+
+        public ListeningExecutorService executor() {
             return cluster.manager.executor;
         }
 
@@ -314,8 +323,8 @@ public class Session {
             return success;
         }
 
-        private Future<?> addOrRenewPool(final Host host) {
-            final HostDistance distance = loadBalancer.distance(host);
+        private ListenableFuture<?> addOrRenewPool(final Host host) {
+            final HostDistance distance = cluster.manager.loadBalancingPolicy().distance(host);
             if (distance == HostDistance.IGNORED)
                 return Futures.immediateFuture(null);
 
@@ -329,19 +338,26 @@ public class Session {
                             previous.shutdown(); // The previous was probably already shutdown but that's ok
                     } catch (AuthenticationException e) {
                         logger.error("Error creating pool to {} ({})", host, e.getMessage());
-                        host.getMonitor().signalConnectionFailure(new ConnectionException(e.getHost(), e.getMessage()));
+                        cluster.manager.signalConnectionFailure(host, new ConnectionException(e.getHost(), e.getMessage()));
                     } catch (ConnectionException e) {
                         logger.debug("Error creating pool to {} ({})", host, e.getMessage());
-                        host.getMonitor().signalConnectionFailure(e);
+                        cluster.manager.signalConnectionFailure(host, e);
                     }
                 }
             });
         }
 
-        private void removePool(Host host) {
-            HostConnectionPool pool = pools.remove(host);
-            if (pool != null)
-                pool.shutdown();
+        ListenableFuture<?> removePool(Host host) {
+            final HostConnectionPool pool = pools.remove(host);
+            if (pool == null)
+                return Futures.immediateFuture(null);
+
+            // Shutdown can take some time and we don't care about holding the thread on that.
+            return executor().submit(new Runnable() {
+                public void run() {
+                    pool.shutdown();
+                }
+            });
         }
 
         /*
@@ -355,11 +371,11 @@ public class Session {
          */
         private void updateCreatedPools() {
             for (Host h : cluster.getMetadata().allHosts()) {
-                HostDistance dist = loadBalancer.distance(h);
+                HostDistance dist = loadBalancingPolicy().distance(h);
                 HostConnectionPool pool = pools.get(h);
 
                 if (pool == null) {
-                    if (dist != HostDistance.IGNORED && h.getMonitor().isUp())
+                    if (dist != HostDistance.IGNORED && h.isUp())
                         addOrRenewPool(h);
                 } else if (dist != pool.hostDistance) {
                     if (dist == HostDistance.IGNORED) {
@@ -373,37 +389,42 @@ public class Session {
 
         @Override
         public void onUp(Host host) {
-            addOrRenewPool(host);
-            loadBalancer.onUp(host);
-            updateCreatedPools();
+            addOrRenewPool(host).addListener(new Runnable() {
+                public void run() {
+                    updateCreatedPools();
+                }
+            }, MoreExecutors.sameThreadExecutor());
         }
 
         @Override
         public void onDown(Host host) {
-            loadBalancer.onDown(host);
             // Note that with well behaved balancing policy (that ignore dead nodes), the removePool call is not necessary
             // since updateCreatedPools should take care of it. But better protect against non well behaving policies.
-            removePool(host);
-            updateCreatedPools();
+            removePool(host).addListener(new Runnable() {
+                public void run() {
+                    updateCreatedPools();
+                }
+            }, MoreExecutors.sameThreadExecutor());
         }
 
         @Override
         public void onAdd(Host host) {
-            addOrRenewPool(host);
-            loadBalancer.onAdd(host);
-            updateCreatedPools();
+            onUp(host);
         }
 
         @Override
         public void onRemove(Host host) {
-            loadBalancer.onRemove(host);
-            removePool(host);
-            updateCreatedPools();
+            onDown(host);
         }
 
         public void setKeyspace(String keyspace) {
+            long timeout = configuration().getSocketOptions().getConnectTimeoutMillis();
             try {
-                Uninterruptibles.getUninterruptibly(executeQuery(new QueryMessage("use " + keyspace, ConsistencyLevel.DEFAULT_CASSANDRA_CL), Query.DEFAULT));
+                Future<?> future = executeQuery(new QueryMessage("use " + keyspace, ConsistencyLevel.DEFAULT_CASSANDRA_CL), Query.DEFAULT);
+                // Note: using the connection timeout is perfectly correct, we should probably change that someday
+                Uninterruptibles.getUninterruptibly(future, timeout, TimeUnit.MILLISECONDS);
+            } catch (TimeoutException e) {
+                throw new DriverInternalError(String.format("No responses after %d milliseconds while setting current keyspace. This should not happen, unless you have setup a very low connection timeout.", timeout));
             } catch (ExecutionException e) {
                 ResultSetFuture.extractCauseFromExecutionException(e);
             }